+++ packages/features/shell/permissions/checkNavigationPermissions.ts
@@ -3,8 +3,8 @@ import { unstable_cache } from "next/cache";
 import { PermissionCheckService } from "@calcom/features/pbac/services/permission-check.service";
 import { MembershipRepository } from "@calcom/lib/server/repository/membership";
 
-import type { NavigationPermissions } from "../NavigationPermissionsProvider";
-import { NAVIGATION_PERMISSION_MAP } from "../NavigationPermissionsProvider";
+import type { NavigationPermissions } from "./types";
+import { NAVIGATION_PERMISSION_MAP } from "./types";
 
 /**
  * Server-side function to check all navigation permissions for a user
@@ -15,6 +15,10 @@ export const checkNavigationPermissions = unstable_cache(
     const permissionService = new PermissionCheckService();
 
     const teamIds = await MembershipRepository.findUserTeamIds({ userId });
+    console.log("💡 teamIds", {
+      teamIds,
+      userId,
+    });
 
     if (teamIds.length === 0) {
       return {
